Split out generic action sheet components
Keep the message specific components separte, so we could re-use the
MenuActionsViewController.
